JSを諸々変更

+++ html/js/home.js
@@ -6,6 +6,7 @@ $(function(){
         dataType:'json',
     })
     .done(function(data){
+        window.taskDataOrg=data;
         window.taskData=data;
         setParentTask(taskData);
         console.log(taskData);
@@ -15,21 +16,23 @@ $(function(){
     function setParentTask(taskData){
         $("#tasks_parent").empty();
         taskData.forEach((element)=>{
-            let rowTask=`
-            <div id="row_parent_${element.id}" class="row-task task-parent" data-id="${element.id}" data-name="${element.name}" data-comment="${element.comment}" onclick="setChildTask(${element.id});setParentComment(${element.id})">
-                <div class="row_task_name">
-                    <input id="parent_${element.id}" type="checkbox" data-id="${element.id}" disabled="disabled" ${(element.check_flag=="1")? 'checked="checked"':''}>
-                    <label for="parent_${element.id}">${element.name}</label>
-                </div>
-                <div class="area-option">
-                    <img class="icon-option" src="/img/option.png" alt="option icon" onclick="toggleParentOption(${element.id})">
-                    <div class="options" id="optionParent${element.id}">
-                        <div data-parent-id="${element.id}" onclick="deleteParentTask(${element.id})">削除</div>
-                        <div data-parent-id="${element.id}" onclick="openModal(1,0,${element.id},null)">編集</div>
+            if(element.delete_flag != 1){
+                let rowTask=`
+                <div id="row_parent_${element.id}" class="row-task task-parent" data-id="${element.id}" data-name="${element.name}" data-comment="${element.comment}" onclick="setChildTask(${element.id});setParentComment(${element.id})">
+                    <div class="row-task-name">
+                        <input id="parent_${element.id}" type="checkbox" data-id="${element.id}" disabled="disabled" ${(element.check_flag == "1")? 'checked="checked"':''}>
+                        <label for="parent_${element.id}">${element.name}</label>
                     </div>
-                </div>
-            </div>`;
-            $("#tasks_parent").append(rowTask);
+                    <div class="area-option">
+                        <img class="icon-option" src="/img/option.png" alt="option icon" onclick="toggleParentOption(${element.id})">
+                        <div class="options" id="optionParent${element.id}">
+                            <div data-parent-id="${element.id}" onclick="deleteParentTask(${element.id})">削除</div>
+                            <div data-parent-id="${element.id}" onclick="openModal(1,0,${element.id},null)">編集</div>
+                        </div>
+                    </div>
+                </div>`;
+                $("#tasks_parent").append(rowTask);
+            }
         });
     }
 
@@ -39,7 +42,7 @@ $(function(){
         // 子タスクを内包する親タスクを検索
         let targetParentTask;
         targetParentTask=taskData.filter((item,index)=>{
-            if(item.id==parentId) return true;
+            if(item.id == parentId) return true;
         })[0];
 
         // 親タスクにフォーカスを当てる
@@ -52,21 +55,23 @@ $(function(){
         // 子タスクを設置
         $("#tasks_child").empty();
         targetParentTask.children.forEach((element)=>{
-            let rowTask=`
-                <div id="row_child_${element.id}" class="row-task" data-parent="${targetParentTask.name}" data-child="${element.name}" data-child-id="${element.id}" data-parent-id="${targetParentTask.id}" data-comment="${element.comment}"  onclick="setChildComment(this)">
-                    <div class="row_task_name">
-                        <input class="child_checkbox" id="child_${element.id}" type="checkbox" ${(element.check_flag=="1")? 'checked="checked"':''} data-id="${element.id}" onchange="childCheckEvent(this.checked,${element.id},${element.parent_id})">
-                        <label for="child_${element.id}">${element.name}</label>
-                    </div>
-                    <div class="area-option">
-                        <img class="icon-option" src="/img/option.png" alt="option icon" onclick="toggleChildOption(${element.id})">
-                        <div class="options" id="optionChild${element.id}">
-                            <div data-parent-id="${element.id}" onclick="deleteChildTask(${targetParentTask.id},${element.id})">削除</div>
-                            <div data-parent-id="${element.id}" onclick="openModal(1,1,${targetParentTask.id},${element.id})">編集</div>
+            if(element.delete_flag != 1){
+                let rowTask=`
+                    <div id="row_child_${element.id}" class="row-task" data-parent="${targetParentTask.name}" data-child="${element.name}" data-child-id="${element.id}" data-parent-id="${targetParentTask.id}" data-comment="${element.comment}"  onclick="setChildComment(this)">
+                        <div class="row-task-name">
+                            <input class="child_checkbox" id="child_${element.id}" type="checkbox" ${(element.check_flag == "1")? 'checked="checked"':''} data-id="${element.id}" onchange="childCheckEvent(this.checked,${element.id},${element.parent_id})">
+                            <label for="child_${element.id}">${element.name}</label>
                         </div>
-                    </div>
-                </div>`;
-            $("#tasks_child").append(rowTask);
+                        <div class="area-option">
+                            <img class="icon-option" src="/img/option.png" alt="option icon" onclick="toggleChildOption(${element.id})">
+                            <div class="options" id="optionChild${element.id}">
+                                <div data-parent-id="${element.id}" onclick="deleteChildTask(${targetParentTask.id},${element.id})">削除</div>
+                                <div data-parent-id="${element.id}" onclick="openModal(1,1,${targetParentTask.id},${element.id})">編集</div>
+                            </div>
+                        </div>
+                    </div>`;
+                $("#tasks_child").append(rowTask);
+            }
         });
 
         // 子タスク作成ボタンに親タスクの情報を埋め込み
@@ -110,11 +115,11 @@ $(function(){
 
     // コメントが変更された時
     window.updateComment=function updateComment(element){
-        if(element.dataset.flag=="0"){
+        if(element.dataset.flag == "0"){
             console.log('parent comment changed');
             updateParentData(element.dataset.id,null,element.value);
         }
-        if(element.dataset.flag=="1"){
+        if(element.dataset.flag == "1"){
             console.log('child comment changed');
             console.log(element);
             updateChildData(element.dataset.childId,element.dataset.parentId,null,null,element.value);
@@ -134,7 +139,7 @@ $(function(){
     function updateParentData(id,name,comment){
         console.log(`${id},${comment}`);
         for(i=0;i<taskData.length;i++){
-            if(taskData[i].id==id){
+            if(taskData[i].id == id){
                 if(name){
                     taskData[i].name=name;
                 }
@@ -149,16 +154,16 @@ $(function(){
         for(i=0; i<taskData.length; i++){
             let numChildren;
             let counter=0;
-            if(taskData[i].id == parentId){
-                numChildren=taskData[i].children.length;
+            if(taskData[i].id  ==  parentId){
+                numChildren=taskData[i].children.filter(function(el){return el.delete_flag != "1"}).length;
                  // 子タスクの内チェックされている物の数を数える
                  taskData[i].children.forEach((element)=>{
-                    if(element.check_flag=="1"){
+                    if(element.delete_flag != "1" && element.check_flag == "1"){
                         counter += 1;
                     }
                 });
                 // 親タスクのcheck_flagを更新
-                if(numChildren==counter){
+                if(numChildren != 0 && numChildren == counter){
                     taskData[i].check_flag="1";
                 } else {
                     taskData[i].check_flag="0";
@@ -175,9 +180,9 @@ $(function(){
     // 子タスクデータのアップデート
     function updateChildData(childId, parentId, isChecked, name,comment){
         taskData.forEach((elParent, indexParent)=>{
-            if(elParent.id==parentId){
+            if(elParent.id == parentId){
                 taskData[indexParent].children.map((elChild,indexChild)=>{
-                    if(elChild.id==childId){
+                    if(elChild.id == childId){
                         if(isChecked!=null){
                             // 子タスクのcheck_flagを更新
                             taskData[indexParent].children[indexChild].check_flag=isChecked ? "1": "0";
@@ -209,27 +214,30 @@ $(function(){
     // タスク追加ボタンをクリック
     window.makeTask=function makeTask(element){
         // 親タスクを追加
-        if(element.dataset.flag==0){
+        if(element.dataset.flag == 0){
             console.log($('#task_name').val());
-            const data={id:countParentNewTask,name:$('#task_name').val(),check_flag:'0',comment:$('#task_comment').val(),del_flag:'0',children:[]};
+            const data={id:countParentNewTask,name:$('#task_name').val(),check_flag:'0',comment:$('#task_comment').val(),delete_flag:'0',children:[]};
             taskData.push(data);
             countParentNewTask--;
             console.log(taskData);
             setParentTask(taskData);
             closeModal();
         // 子タスクを追加
-        }else if(element.dataset.flag==1){
-            const data={id:countChildNewTask,name:$('#task_name').val(),parent_id:element.dataset.parentId,check_flag:'0',comment:$('#task_comment').val(),del_flag:'0'};
+        }else if(element.dataset.flag == 1){
+            const data={id:countChildNewTask,name:$('#task_name').val(),parent_id:element.dataset.parentId,check_flag:'0',comment:$('#task_comment').val(),delete_flag:'0'};
             for(i=0;i<taskData.length;i++){
-                if(taskData[i].id==element.dataset.parentId){
+                if(taskData[i].id == element.dataset.parentId){
                     // 親タスクに新規作成した子タスクを追加
                     taskData[i].children.push(data);
                     countChildNewTask--;
                     console.log(taskData);
+                    // 親タスクにチェックをつけるか？
+                    checkParent(element.dataset.parentId);
+                    setParentTask(taskData);
                     setChildTask(element.dataset.parentId);
                     closeModal();
-                    break;
                 }
+                    break;
             }
 
         }
@@ -238,12 +246,12 @@ $(function(){
     // タスク編集ボタンをクリック
     window.editTask=function editTask(element){
         // 親タスクを編集
-        if(element.dataset.flag==0){
+        if(element.dataset.flag == 0){
             updateParentData(element.dataset.parentId,$('#task_name').val(),$('#task_comment').val());
             setParentTask(taskData);
             closeModal();
         // 子タスクを追加
-        }else if(element.dataset.flag==1){
+        }else if(element.dataset.flag == 1){
             updateChildData(element.dataset.parentId,element.dataset.childId,null,$('#task_name').val(),$('#task_comment').val());
             closeModal();
         }
@@ -252,16 +260,16 @@ $(function(){
     // データ抽出
     function getParentRecord(id){
         for(i=0;i<taskData.length;i++){
-            if(taskData[i].id==id){
+            if(taskData[i].id == id){
                 return taskData[i];
             }
         }
     }
     function getChildRecord(parentId,childId){
         for(ip=0; ip<taskData.length;ip++){
-            if(taskData[ip].id==parentId){
+            if(taskData[ip].id == parentId){
                 for(ic=0; ic<taskData[ip].children.length; ic++){
-                    if(taskData[ip].children[ic].id==childId){
+                    if(taskData[ip].children[ic].id == childId){
                         return taskData[ip].children[ic];
                     }
                 }
@@ -286,28 +294,33 @@ $(function(){
     // タスク削除
     window.deleteParentTask=function deleteParentTask(parentId){
         for(i=0; i<taskData.length; i++){
-            if(taskData[i].id==parentId){
-                taskData.splice(i,1);
+            if(taskData[i].id == parentId){
+                taskData[i].delete_flag = "1";
+                // 内包する子タスク全てを削除
+                for(ic=0; ic<taskData[i].children.length; ic++){
+                    taskData[i].children[ic].delete_flag="1";
+                }
                 break;
             }
         }
-        console.log(taskData);
-        setParentTask(taskData);
         $("#tasks_child").empty();
-
+        setParentTask(taskData);
+        console.log(taskData);
     }
     window.deleteChildTask=function deleteChildTask(parentId,childId){
         for(pi=0; pi<taskData.length; pi++){
-            if(taskData[pi].id==parentId){
+            if(taskData[pi].id == parentId){
                 for(ci=0; ci<taskData[pi].children.length; ci++ ){
-                    if(taskData[pi].children[ci].id==childId){
-                        taskData[pi].children.splice(ci,1);
+                    if(taskData[pi].children[ci].id == childId){
+                        taskData[pi].children[ci].delete_flag = 1;
                         break;
                     }
                 }
                 break;
             }
         }
+        checkParent(parentId);
+        setParentTask(taskData);
         setChildTask(parentId);
 
     }
@@ -315,13 +328,13 @@ $(function(){
     // タスク作成,編集モーダルを開く openModal(追加、編集を判別する,親タスク、子タスクを判別,parentId,childId)
     window.openModal=function openModal(typeFlag,flag,parentId,childId){
         // 新規タスク追加
-        if(typeFlag==0){
+        if(typeFlag == 0){
             console.log(parentId);
             // 親タスクを作成するモーダル
-            if(flag==0){
+            if(flag == 0){
                 $('#task_name').val('');
                 $('#task_comment').val('');
-                $(".make-task-header").text('');
+                $(".make-task-header").text('新規親タスク作成');
                 $("#modal_overlay").css('display','block');
                 $("#modal_make_task").css('display','block');
                 $('#task_name').val('');
@@ -334,7 +347,7 @@ $(function(){
                 });
     
             // 子タスクを作成するモーダル
-            }else if(flag==1){
+            }else if(flag == 1){
                 if(parentId){
                     const data=getParentRecord(parentId);
                     console.log(data);
@@ -357,10 +370,10 @@ $(function(){
             }
         }
         // タスク編集用モーダル
-        if(typeFlag==1){
+        if(typeFlag == 1){
             console.log(parentId);
             // 親タスクを編集するモーダル
-            if(flag==0){
+            if(flag == 0){
                 const data=getParentRecord(parentId);
                 $('#task_name').val(data.name);
                 $('#task_comment').val(data.comment);
@@ -376,7 +389,7 @@ $(function(){
                 });
 
             // 子タスクを編集するモーダル
-            }else if(flag==1){
+            }else if(flag == 1){
                 if(parentId){
                     const data=getChildRecord(parentId,childId);
                     console.log(data);
@@ -414,15 +427,17 @@ $(function(){
     // データをバックエンドに送信
     window.submitData= function submitData(){
         $.ajax({
-            url: "/dataStore",
+            url: "/home/dataStore",
             type:'POST',
             dataType: 'json',
-            data : taskData,
-            timeout:3000,
+            data: {taskDataNew: taskData, taskDataOrg: taskDataOrg},
         }).done(function(data) {
-                          alert("ok");
+            alert('全ての変更が保存されました。');
+            console.log(data);
+            location.reload();
         }).fail(function(XMLHttpRequest, textStatus, errorThrown) {
-                         alert("error");
+            console.log(errorThrown);
+            alert('変更の保存に失敗しました。');
         })
     }
